Replace some hardcoded text, minor refactoring
